Change local var type to size_t

+++ src/options.c
@@ -15,7 +15,7 @@
 #include <yaz/options.h>
 
 static int arg_no = 1;
-static int arg_off = 0;
+static size_t arg_off = 0;
 
 int options (const char *desc, char **argv, int argc, char **arg)
 {
